pyproject.toml
[project]
name = "aicapture"
version = "0.3.6"
description = "A Python module to capture knowledge from documents using Vision Language Models (VLMs)"
authors = [
    {name = "Aitomatic"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["vision", "language", "models", "vlm", "document", "capture", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

# Runtime dependencies
dependencies = [
    "pymupdf>=1.25.3",  # PDF processing and manipulation library
    "pillow>=11.1.0",  # Python Imaging Library for image processing
    "anthropic>=0.30.1",  # Anthropic's Claude API client
    "openai>=2.0.0",  # OpenAI API client for GPT models
    "python-dotenv>=1.0.1",  # Load environment variables from .env files
    "boto3>=1.34.0",  # AWS SDK for Python
    "loguru>=0.7.3",  # Better logging library than Python's built-in logging
    "numpy>=2.2.3",  # Numerical computing library for array operations
    "opencv-python>=4.11.0.86",  # Computer vision library
]

[project.optional-dependencies]
# Development dependencies
dev = [
    "pytest>=8.0.0",  # Testing framework
    "pytest-asyncio>=0.23.5",  # Async testing support for pytest
    "pytest-cov>=4.1.0",  # Test coverage reporting
    "ruff>=0.8.0",  # Fast Python linter and formatter (replaces black, isort, flake8)
    "mypy>=1.8.0",  # Static type checker
    "pre-commit>=3.6.0",  # Git hooks manager
]
# Optional cloud provider integrations
aws = ["boto3>=1.34.0"]  # Optional AWS integration
google = ["google-cloud-vision"]  # Optional Google Cloud Vision integration
azure = ["azure-cognitiveservices-vision-computervision"]  # Optional Azure Computer Vision integration

[project.urls]
Homepage = "https://github.com/Aitomatic/ai-vision-capture"
Repository = "https://github.com/Aitomatic/ai-vision-capture"
Issues = "https://github.com/Aitomatic/ai-vision-capture/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
]
ignore = [
    "E203",  # whitespace before ':'
    "E501",  # line too long (handled by formatter)
    "UP",    # pyupgrade (ignore Python version upgrade suggestions)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Unused imports in __init__.py
"tests/*" = ["E501"]  # Long lines in tests

[tool.ruff.lint.isort]
known-first-party = ["aicapture"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
check_untyped_defs = false
disallow_any_generics = false
disallow_untyped_calls = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=aicapture --cov-report=term-missing"
asyncio_mode = "auto"
